third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/characteristics-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getTwoMeasurementIntervalCharacteristicsService()
    .then(({service}) => service.getCharacteristics())
    .then(characteristics => {
      // Expect four characteristic instances.
      assert_equals(characteristics.length, 4);

      let uuid_set = new Set(characteristics.map(c => c.uuid));
      // Two of the expected characteristics are 'measurement_interval',
      // so only 3 unique UUIDs.
      assert_equals(uuid_set.size, 3);

      assert_true(uuid_set.has(BluetoothUUID.getCharacteristic(
        'temperature_measurement')));
      assert_true(uuid_set.has(BluetoothUUID.getCharacteristic(
        'temperature_type')));
      assert_true(uuid_set.has(BluetoothUUID.getCharacteristic(
        'measurement_interval')));
    });
}, 'Find all characteristics in a service.');
</script>
